+++ database/migrations/2018_01_11_125205_create_wallets_table.php
@@ -17,7 +17,7 @@ class CreateWalletsTable extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('user_id');
-            $table->integer('balance');
+            $table->bigInteger('balance');
             $table->timestamps();
         });
     }
